- 0.4.3 : Drag & drop skeleton

+++ videomanager-api/src/main/java/com/gamergeo/project/videomanager/persistence/VideoDao.java
@@ -7,9 +7,10 @@ import org.springframework.data.repository.CrudRepository;
 import com.gamergeo.project.videomanager.model.Video;
 
 //@Repository("videoDao")
-public interface VideoDao extends CrudRepository<Video, Integer> {
+public interface VideoDao extends CrudRepository<Video, Long> {
 	
 	List<Video> findAll();
 	
 	List<Video> findByTitleContaining(String title);
+	
 }
